Making the minutes show

+++ src/main/groovy/com/meetupinthemiddle/model/Person.groovy
@@ -11,7 +11,7 @@ class Person {
   @NotNull
   String name
   float distance
-  int travelTime
+  long travelTime
   @NotNull
   String from
   LatLong latLong
